requirements.txt
qrcode[pil]==7.4.2
PyYAML==6.0.1
Pillow==10.3.0
